posts/DoingDataWranglingRight/requirements.txt
pandas==1.5.3
plotly-express==0.4.1
sweetviz==2.2.1
